# openapi.yml
openapi: 3.0.3
info:
  description: API for managing Sinopia resources atop Mongo / AWS DocumentDB platform
  version: 1.0.0
  title: Sinopia API
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://{environment}.sinopia.io"
    variables:
      environment:
        default: api
        description: Deployment environments
        enum:
          - api
          - api.stage
          - api.development
...
paths:
...
  /marc/{resourceId}/job/{username}/{timestamp}:
    get:
      summary: Returns the status of the MARC conversion job
      tags:
        - marc
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
        "303":
          description: Redirect to MARC version url
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: Server error indicating the conversion failed
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/Errors"
      parameters:
        - name: resourceId
          in: path
          description: ID of the resource
          required: true
          schema:
            type: string
        - name: username
          in: path
          description: Username of the user that requested the MARC conversion job
          required: true
          schema:
            type: string
        - name: timestamp
          in: path
          description: Timestamp (version) of the resource to retrieve
          required: true
          schema:
            type: string
        - name: accept
          in: header
          schema:
            type: string
  /marc/{resourceId}/version/{username}/{timestamp}:
    get:
      summary: Return the marc record for the given resource version
      tags:
        - marc
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
            application/marc:
              schema:
                type: string
      parameters:
        - name: resourceId
          in: path
          description: ID of the resource
          required: true
          schema:
            type: string
        - name: username
          in: path
          description: Username of the resource creator
          required: true
          schema:
            type: string
        - name: timestamp
          in: path
          description: Timestamp (version) of the resource to retrieve
          required: true
          schema:
            type: string
  /marc/{resourceId}:
    post:
      summary: Request conversion of a resource to MARC
      security:
        - bearerAuth: []
      tags:
        - marc
      responses:
        "202":
          description: Accepted
          headers:
            content-location:
              description: The URL for checking the status of the MARC conversion job.
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
      parameters:
        - name: resourceId
          in: path
          description: ID of the resource
          required: true
          schema:
            type: string
...
components:
  schemas:
...
    Errors:
      description: Errors response.
      type: array
      items:
        type: object
        properties:
          title:
            type: string
          details:
            type: string
          status:
            type: string
        required:
          - title
          - status
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT